# Fix data races in sorting and Reverse