--- a/hugolib/pageSort.go
+++ b/hugolib/pageSort.go
@@ -17,6 +17,8 @@ import (
 	"sort"
 )
 
+var spc = newPageCache()
+
 /*
  * Implementation of a custom sorter for Pages
  */
@@ -62,60 +64,121 @@ func (p Pages) Limit(n int) Pages {
 	return p
 }
 
+// ByWeight sorts the Pages by weight and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByWeight() Pages {
-	PageBy(DefaultPageSort).Sort(p)
-	return p
+	key := "pageSort.ByWeight"
+	pages, _ := spc.get(key, p, PageBy(DefaultPageSort).Sort)
+	return pages
 }
 
+// ByTitle sorts the Pages by title and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByTitle() Pages {
+
+	key := "pageSort.ByTitle"
+
 	title := func(p1, p2 *Page) bool {
 		return p1.Title < p2.Title
 	}
 
-	PageBy(title).Sort(p)
-	return p
+	pages, _ := spc.get(key, p, PageBy(title).Sort)
+	return pages
 }
 
+// ByLinkTitle sorts the Pages by link title and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByLinkTitle() Pages {
+
+	key := "pageSort.ByLinkTitle"
+
 	linkTitle := func(p1, p2 *Page) bool {
 		return p1.linkTitle < p2.linkTitle
 	}
 
-	PageBy(linkTitle).Sort(p)
-	return p
+	pages, _ := spc.get(key, p, PageBy(linkTitle).Sort)
+
+	return pages
 }
 
+// ByDate sorts the Pages by date and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByDate() Pages {
+
+	key := "pageSort.ByDate"
+
 	date := func(p1, p2 *Page) bool {
 		return p1.Date.Unix() < p2.Date.Unix()
 	}
 
-	PageBy(date).Sort(p)
-	return p
+	pages, _ := spc.get(key, p, PageBy(date).Sort)
+
+	return pages
 }
 
+// ByPublishDate sorts the Pages by publish date and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByPublishDate() Pages {
+
+	key := "pageSort.ByPublishDate"
+
 	pubDate := func(p1, p2 *Page) bool {
 		return p1.PublishDate.Unix() < p2.PublishDate.Unix()
 	}
 
-	PageBy(pubDate).Sort(p)
-	return p
+	pages, _ := spc.get(key, p, PageBy(pubDate).Sort)
+
+	return pages
 }
 
+// ByLength sorts the Pages by length and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) ByLength() Pages {
+
+	key := "pageSort.ByLength"
+
 	length := func(p1, p2 *Page) bool {
 		return len(p1.Content) < len(p2.Content)
 	}
 
-	PageBy(length).Sort(p)
-	return p
+	pages, _ := spc.get(key, p, PageBy(length).Sort)
+
+	return pages
 }
 
+// Reverse reverses the order in Pages and returns a copy.
+//
+// Adjacent invocactions on the same receiver will return a cached result.
+//
+// This may safely be executed  in parallel.
 func (p Pages) Reverse() Pages {
-	for i, j := 0, len(p)-1; i < j; i, j = i+1, j-1 {
-		p[i], p[j] = p[j], p[i]
+	key := "pageSort.Reverse"
+
+	reverseFunc := func(pages Pages) {
+		for i, j := 0, len(pages)-1; i < j; i, j = i+1, j-1 {
+			pages[i], pages[j] = pages[j], pages[i]
+		}
 	}
 
-	return p
+	pages, _ := spc.get(key, p, reverseFunc)
+
+	return pages
 }