@bep 2015-07-21
Of course, just making copies on every sort and reverse will also fix the main issue here, but at a slight cost:

```
Your branch is up-to-date with 'origin/master'.
benchmark                           old ns/op     new ns/op     delta
BenchmarkSortByWeightAndReverse     14304         19858         +38.83%

benchmark                           old allocs     new allocs     delta
BenchmarkSortByWeightAndReverse     1              3              +200.00%

benchmark                           old bytes     new bytes     delta
BenchmarkSortByWeightAndReverse     32            5152          +16000.00%
```

@tatsushid @spf13 ... any thoughts?