@spf13 2015-07-22
wow. Thanks for tracking this down. I think caching is a big win, especially for large sites. I'm not sure I fully understand the root issue. Is it that the sort / reverse isn't safe for concurrency and consequently the order isn't right/ items get lost?

It seems the best approach would be to make it concurrently safe (mutexes or channels would work well). Then cache the results.

Is this what you were looking for?