@bep 2015-07-22
There are several issues in play here:
* The symptom in this case is "lost pages" caused by a data race where `Reverse` modified the slice while `ByDate`read it.
* There are lots of functions that operate on `Pages` that does restructuring of the slice. 
* Most use cases use this modified slice only once (in the index template)
* In this case, the single template had a list of recent posts in the side-bar, so the `Data.Pages.ByDate.Reverse` was executed 300 times in parallell => data race.
* Then there is, of course, variants of the above, people trying do multiple sorting of the same data etc.

My fix is this:

* A copy is made of the slice before doing any structural changes. This fixes the concurrency issues / data races etc.
* In addition to the above, the result of the restructuring is cached. In normal cases (small blogs), there is very minute savings from cache this result, but it just sound ineffective to sort and reverse the slice 300 times with the same result.